pyproject.toml
[project]
name = "livrariaTutorial"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "GuilhermeSchreiber7", email = "guilhermeschreieber2007@gmail.com"},
]
dependencies = ["django>=5.2", "djangorestframework>=3.16.0", "django-cors-headers>=4.7.0", "drf-spectacular>=0.28.0"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
